Don't let Debian (& friends) compinit
The zshrc file provided by Debian (and related distributions) at
/etc/zsh/zshrc runs compinit.  This means if we run compinit in our own
dotfiles, we will run it twice with minor impact to our startup time.
It also means if we are lazy and add a completion path to our user zshrc
(instead of user zshenv), compinit will run before we modify fpath
causing a complete rebuild and dump of completions since the file count
will change with significant impact to our startup time.  So let's
follow the instructions in the system zshrc file:

 If you don't want compinit called here, place the line
 skip_global_compinit=1
 in your $ZDOTDIR/.zshenv